Возвращение утраченного функционала SDK API 1.0 [Review]  - Исправлена документация над полями RateEntity rate  - У части полкй created тип изменён на корректный LocalDateTime  - У части денежных полей тип изменён на корректный Long  - У полей сущностей типа ListEntity проставлены классы дженериков  - У некоторых полей сущностей типа MetaEntity проставлены конкретные классы сущностей

+++ src/main/java/com/lognex/api/entities/documents/RetailDemandDocumentEntity.java
@@ -27,7 +27,7 @@ public class RetailDemandDocumentEntity extends DocumentEntity {
     private Boolean fiscal;
     private Long noCashSum;
     private OrganizationEntity organization;
-    private Integer payedSum;
+    private Long payedSum;
     private ListEntity<DocumentPosition> positions;
     private RateEntity rate;
     private RetailShiftEntity retailShift;
